Implement updates to allow switching WAN/LAN interfaces.
This patch implements a frontend option to allow switching WAN/LAN interfaces
when routers have two. This allows several routers to be used as switches for
the uplink network, in case they have a switch for one of the interfaces,
and not just as a set of Mesh/LAN ports.