Import patchset for implementing allow_remote and some adaptations in wireless config.
This patchset lays the groundwork for implementing allow_remote, which sends a flag to
the gateway through the Alfred protocol to signal whether the node wants to have a firewall
set up so that it is not accessible by its public IPv6 address(es) from any routed (and
not bridged) network.

Additionally, it adds the configurable possibility to set up a new, additional client
network for the Freifunk client network to allow merging of different locations but
keeping the old names (specific to a certain community) intact.

Finally, this removes the broken flag on the ramips build to allow building a testing
binary for the D-LINK DIR-615 D-series, which should work with Gluon.